Eliminate panics from live code (#73)
* Eliminate panics from live code.

As a library, panicking, even in the face of a bug, is a problem.

Also, a lot of warnings (mostly `try!` vs `?` and `dyn`) were removed,
as part of the edit/rebuild/test cycle.

* Update rust dependencies to rust 1.46

If Slackware thinks Rust 1.46 is stable enough for everyday use (i.e.,
building Mozilla) then it should be stable enough for anyone who doesn't
track "stable".

* Prevent possible infinite recursion

`cargo clippy` warned of infinite recursion, so let's heed the warning

Co-authored-by: Ben Stern <bstern@fortian.com>